@@ -7,9 +7,9 @@ namespace MaterialDesignDemo.Converters
 {
     public class BrushToHexConverter : IValueConverter
     {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
         {
-            if (value == null) return null;
+            if (value is null) return null;
             string lowerHexString(int i) => i.ToString("X2").ToLower();
             var brush = (SolidColorBrush)value;
             var hex = lowerHexString(brush.Color.R) +
@@ -18,9 +18,7 @@ namespace MaterialDesignDemo.Converters
             return "#" + hex;
         }
 
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
+        public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
+            => throw new NotImplementedException();
     }
 }
