@@ -11,27 +11,21 @@ namespace MaterialDesignDemo.Domain
         private readonly Action<object> _execute;
         private readonly Func<object, bool> _canExecute;
 
-        public AnotherCommandImplementation(Action<object> execute) : this(execute, null)
-        {
-        }
+        public AnotherCommandImplementation(Action<object> execute)
+            : this(execute, null)
+        { }
 
-        public AnotherCommandImplementation(Action<object> execute, Func<object, bool> canExecute)
+        public AnotherCommandImplementation(Action<object> execute, Func<object, bool>? canExecute)
         {
-            if (execute == null) throw new ArgumentNullException(nameof(execute));
+            if (execute is null) throw new ArgumentNullException(nameof(execute));
 
             _execute = execute;
             _canExecute = canExecute ?? (x => true);
         }
 
-        public bool CanExecute(object parameter)
-        {
-            return _canExecute(parameter);
-        }
+        public bool CanExecute(object parameter) => _canExecute(parameter);
 
-        public void Execute(object parameter)
-        {
-            _execute(parameter);
-        }
+        public void Execute(object parameter) => _execute(parameter);
 
         public event EventHandler CanExecuteChanged
         {
@@ -45,9 +39,6 @@ namespace MaterialDesignDemo.Domain
             }
         }
 
-        public void Refresh()
-        {
-            CommandManager.InvalidateRequerySuggested();
-        }
+        public void Refresh() => CommandManager.InvalidateRequerySuggested();
     }
 }
