@@ -7,10 +7,10 @@ namespace MaterialDesignDemo
         public Chips() => InitializeComponent();
 
         private void ButtonsDemoChip_OnClick(object sender, RoutedEventArgs e)
-            => MainWindow.Snackbar.MessageQueue.Enqueue("Chip clicked!");
+            => MainWindow.Snackbar.MessageQueue?.Enqueue("Chip clicked!");
 
         private void ButtonsDemoChip_OnDeleteClick(object sender, RoutedEventArgs e)
-            => MainWindow.Snackbar.MessageQueue.Enqueue("Chip delete clicked!");
+            => MainWindow.Snackbar.MessageQueue?.Enqueue("Chip delete clicked!");
 
     }
 }
