@@ -3,19 +3,18 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
-using MaterialDesignDemo.Domain;
 
 namespace MaterialDesignDemo.Domain
 {
     public class DemoItem : INotifyPropertyChanged
     {
         private string _name;
-        private object _content;
+        private object? _content;
         private ScrollBarVisibility _horizontalScrollBarVisibilityRequirement;
         private ScrollBarVisibility _verticalScrollBarVisibilityRequirement;
         private Thickness _marginRequirement = new Thickness(16);
 
-        public DemoItem(string name, object content, IEnumerable<DocumentationLink> documentation)
+        public DemoItem(string name, object? content, IEnumerable<DocumentationLink> documentation)
         {
             _name = name;
             Content = content;
@@ -24,41 +23,39 @@ namespace MaterialDesignDemo.Domain
 
         public string Name
         {
-            get { return _name; }
-            set { this.MutateVerbose(ref _name, value, RaisePropertyChanged()); }
+            get => _name;
+            set => this.MutateVerbose(ref _name, value, RaisePropertyChanged());
         }
 
-        public object Content
+        public object? Content
         {
-            get { return _content; }
-            set { this.MutateVerbose(ref _content, value, RaisePropertyChanged()); }
+            get => _content;
+            set => this.MutateVerbose(ref _content, value, RaisePropertyChanged());
         }
 
         public ScrollBarVisibility HorizontalScrollBarVisibilityRequirement
         {
-            get { return _horizontalScrollBarVisibilityRequirement; }
-            set { this.MutateVerbose(ref _horizontalScrollBarVisibilityRequirement, value, RaisePropertyChanged()); }
+            get => _horizontalScrollBarVisibilityRequirement;
+            set => this.MutateVerbose(ref _horizontalScrollBarVisibilityRequirement, value, RaisePropertyChanged());
         }
 
         public ScrollBarVisibility VerticalScrollBarVisibilityRequirement
         {
-            get { return _verticalScrollBarVisibilityRequirement; }
-            set { this.MutateVerbose(ref _verticalScrollBarVisibilityRequirement, value, RaisePropertyChanged()); }
+            get => _verticalScrollBarVisibilityRequirement;
+            set => this.MutateVerbose(ref _verticalScrollBarVisibilityRequirement, value, RaisePropertyChanged());
         }
 
         public Thickness MarginRequirement
         {
-            get { return _marginRequirement; }
-            set { this.MutateVerbose(ref _marginRequirement, value, RaisePropertyChanged()); }
+            get => _marginRequirement;
+            set => this.MutateVerbose(ref _marginRequirement, value, RaisePropertyChanged());
         }
 
         public IEnumerable<DocumentationLink> Documentation { get; }
 
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         private Action<PropertyChangedEventArgs> RaisePropertyChanged()
-        {
-            return args => PropertyChanged?.Invoke(this, args);
-        }
+            => args => PropertyChanged?.Invoke(this, args);
     }
 }
