@@ -40,9 +40,7 @@ namespace MaterialDesignDemo.Domain
         }
 
         private void ClosingEventHandler(object sender, DialogClosingEventArgs eventArgs)
-        {
-            Console.WriteLine("You can intercept the closing event, and cancel here.");
-        }
+            => Console.WriteLine("You can intercept the closing event, and cancel here.");
 
         private async void ExecuteRunExtendedDialog(object o)
         {
@@ -60,13 +58,12 @@ namespace MaterialDesignDemo.Domain
         }
 
         private void ExtendedOpenedEventHandler(object sender, DialogOpenedEventArgs eventargs)
-        {
-            Console.WriteLine("You could intercept the open and affect the dialog using eventArgs.Session.");
-        }
+            => Console.WriteLine("You could intercept the open and affect the dialog using eventArgs.Session.");
 
         private void ExtendedClosingEventHandler(object sender, DialogClosingEventArgs eventArgs)
         {
-            if ((bool)eventArgs.Parameter == false) return;
+            if (eventArgs.Parameter is bool parameter &&
+                parameter == false) return;
 
             //OK, lets cancel the close...
             eventArgs.Cancel();
@@ -91,11 +88,11 @@ namespace MaterialDesignDemo.Domain
         public ICommand CancelSample4DialogCommand { get; }
 
         private bool _isSample4DialogOpen;
-        private object _sample4Content;
+        private object? _sample4Content;
 
         public bool IsSample4DialogOpen
         {
-            get { return _isSample4DialogOpen; }
+            get => _isSample4DialogOpen;
             set
             {
                 if (_isSample4DialogOpen == value) return;
@@ -104,9 +101,9 @@ namespace MaterialDesignDemo.Domain
             }
         }
 
-        public object Sample4Content
+        public object? Sample4Content
         {
-            get { return _sample4Content; }
+            get => _sample4Content;
             set
             {
                 if (_sample4Content == value) return;
@@ -137,7 +134,7 @@ namespace MaterialDesignDemo.Domain
 
         #endregion
 
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
         {
