@@ -6,22 +6,17 @@ namespace MaterialDesignDemo.Domain
 {
     public class SampleDialogViewModel : INotifyPropertyChanged
     {
-        private string _name;
+        private string? _name;
 
-        public string Name
+        public string? Name
         {
-            get { return _name; }
-            set
-            {
-                this.MutateVerbose(ref _name, value, RaisePropertyChanged());
-            }
+            get => _name;
+            set => this.MutateVerbose(ref _name, value, RaisePropertyChanged());
         }
 
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         private Action<PropertyChangedEventArgs> RaisePropertyChanged()
-        {
-            return args => PropertyChanged?.Invoke(this, args);
-        }
+            => args => PropertyChanged?.Invoke(this, args);
     }
 }
