@@ -5,9 +5,9 @@ namespace MaterialDesignDemo.Domain
 {
     public class SimpleDataTemplateSelector : DataTemplateSelector
     {
-        public DataTemplate FixedTemplate { get; set; }
+        public DataTemplate? FixedTemplate { get; set; }
 
-        public override DataTemplate SelectTemplate(object item, DependencyObject container)
+        public override DataTemplate? SelectTemplate(object item, DependencyObject container)
         {
             return FixedTemplate;
         }
