@@ -8,10 +8,10 @@ namespace MaterialDesignDemo.Domain
 {
     public class FieldsViewModel : INotifyPropertyChanged
     {
-        private string _name;
-        private string _name2;
+        private string? _name;
+        private string? _name2;
         private int? _selectedValueOne;
-        private string _selectedTextTwo;
+        private string? _selectedTextTwo;
 
         public FieldsViewModel()
         {
@@ -21,51 +21,36 @@ namespace MaterialDesignDemo.Domain
             SelectedTextTwo = null;
         }
 
-        public string Name
+        public string? Name
         {
-            get { return _name; }
-            set
-            {
-                this.MutateVerbose(ref _name, value, RaisePropertyChanged());
-            }
+            get => _name;
+            set => this.MutateVerbose(ref _name, value, RaisePropertyChanged());
         }
 
-        public string Name2
+        public string? Name2
         {
-            get { return _name2; }
-            set
-            {
-                this.MutateVerbose(ref _name2, value, RaisePropertyChanged());
-            }
+            get => _name2;
+            set => this.MutateVerbose(ref _name2, value, RaisePropertyChanged());
         }
 
         public int? SelectedValueOne
         {
-            get { return _selectedValueOne; }
-            set
-            {
-                this.MutateVerbose(ref _selectedValueOne, value, RaisePropertyChanged());
-            }
+            get => _selectedValueOne;
+            set => this.MutateVerbose(ref _selectedValueOne, value, RaisePropertyChanged());
         }
 
-        public string SelectedTextTwo
+        public string? SelectedTextTwo
         {
-            get { return _selectedTextTwo; }
-            set
-            {
-                this.MutateVerbose(ref _selectedTextTwo, value, RaisePropertyChanged());
-            }
+            get => _selectedTextTwo;
+            set => this.MutateVerbose(ref _selectedTextTwo, value, RaisePropertyChanged());
         }
 
         public IList<int> LongListToTestComboVirtualization { get; }
 
         public DemoItem DemoItem => new DemoItem("Mr. Test", null, Enumerable.Empty<DocumentationLink>());
 
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
-        private Action<PropertyChangedEventArgs> RaisePropertyChanged()
-        {
-            return args => PropertyChanged?.Invoke(this, args);
-        }
+        private Action<PropertyChangedEventArgs> RaisePropertyChanged() => args => PropertyChanged?.Invoke(this, args);
     }
 }
